Data Science/Conversational RAG Chatbot/requirements.txt
langchain
langchain-core
langchain_community
docx2txt
pypdf
langchain_chroma
python-multipart
fastapi
uvicorn
langchain-groq
python-dotenv
streamlit
beautifulsoup4
unstructured
langchain-huggingface
pydantic




